✨ feat(Dashboard): restructure components for better organization, add new Sidebar, OrderTable, and ProductDetails components, and implement dynamic filtering and pagination for enhanced user experience
