+++ src/components/dashboard/ProductDetails.tsx
